chore: moved plugins under seperate directory
