+++ src/module.ts
@@ -6,13 +6,13 @@ import type { AssetURLOptions } from 'vue/compiler-sfc'
 import type { Options as SassOptions } from 'sass'
 import { version } from '../package.json'
 import { vuePluginTemplate } from './plugin'
-import { transformDirectivesPlugin } from './transform/directives'
+import { transformDirectivesPlugin } from './plugins/transform/directives'
 import type { ModuleContext, QuasarFontIconSets, QuasarFrameworkInnerConfiguration, QuasarImports, QuasarPlugins, QuasarSvgIconSets } from './types'
-import { transformScssPlugin } from './transform/scss'
+import { transformScssPlugin } from './plugins/transform/scss'
 import { importJSON, kebabCase } from './utils'
-import { virtualQuasarEntryPlugin } from './virtual/entry'
-import { virtualAnimationsPlugin } from './virtual/animations'
-import { virtualBrandPlugin } from './virtual/brand'
+import { virtualQuasarEntryPlugin } from './plugins/virtual/entry'
+import { virtualAnimationsPlugin } from './plugins/virtual/animations'
+import { virtualBrandPlugin } from './plugins/virtual/brand'
 import { resolveFont, resolveFontIcon } from './resolve'
 import { quasarAnimationsPath, quasarBrandPath, quasarCssPath, quasarFontsPath, quasarIconsPath } from './constants'
 import { transformImportPlugin } from './transform/import'
