+++ src/plugins/transform/import.ts
@@ -1,6 +1,6 @@
 import { createFilter } from '@rollup/pluginutils'
 import { createUnplugin } from 'unplugin'
-import type { ModuleContext } from '../types'
+import type { ModuleContext } from '../../types'
 
 interface VueQuery {
   vue?: boolean
