+++ src/plugins/virtual/animations.ts
@@ -1,9 +1,9 @@
 import { createUnplugin } from 'unplugin'
 import { logger } from '@nuxt/kit'
-import { allAnimationValues, quasarAnimationsPath } from '../constants'
-import { resolveAnimation } from '../resolve'
-import type { ModuleContext } from '../types'
-import { readFileMemoized } from '../utils'
+import { allAnimationValues, quasarAnimationsPath } from '../../constants'
+import { resolveAnimation } from '../../resolve'
+import type { ModuleContext } from '../../types'
+import { readFileMemoized } from '../../utils'
 
 // Add css suffix so loaded string can be interpreted as a css file
 const RESOLVED_ID = '/quasar/animations.css'
