+++ src/plugins/virtual/brand.ts
@@ -1,6 +1,6 @@
 import { createUnplugin } from 'unplugin'
-import { quasarBrandPath } from '../constants'
-import type { ModuleContext } from '../types'
+import { quasarBrandPath } from '../../constants'
+import type { ModuleContext } from '../../types'
 
 // Add css suffix so loaded string can be interpreted as a css file
 const RESOLVED_ID = '/quasar/brand.css'
