+++ src/plugins/virtual/entry.ts
@@ -1,5 +1,5 @@
 import { createUnplugin } from 'unplugin'
-import { importJSON } from '../utils'
+import { importJSON } from '../../utils'
 
 const QUASAR_ENTRY = 'quasar'
 const QUASAR_VIRTUAL_ENTRY = `\0${QUASAR_ENTRY}`
